fix(BrowserAi): remove unsafe Float32Array cast in ONNX worker
TensorData.data already accepts Float32Array | BigInt64Array |
Int32Array | Uint8Array — the `as Float32Array` cast was hiding
the actual type from callers.

+++ src/modules/BrowserAi/workers/onnxInferenceWorker.ts
@@ -387,7 +387,7 @@ self.onmessage = async (event: MessageEvent<WorkerRequest>): Promise<void> => {
                 } else if (val.data instanceof Uint8Array) {
                     dataType = 'bool';
                 }
-                resultOutputs[key] = { data: val.data as Float32Array, dims: val.dims, type: dataType };
+                resultOutputs[key] = { data: val.data, dims: val.dims, type: dataType };
             }
             const response: WorkerResponse = { type: 'inference-result', requestId: req.requestId, outputs: resultOutputs };
             self.postMessage(response);
